backend-django/requirements.txt
# Django Backend Requirements
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0

# RAG & ML
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
openai>=1.0.0
tiktoken>=0.5.0

# Document Processing
pypdf>=3.17.0
nltk>=3.8.0

# Optional: LangChain & LlamaIndex
langchain>=0.1.0
langchain-community>=0.0.10
llama-index>=0.9.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL (optional, can use SQLite)

# Utilities
httpx>=0.25.0
pydantic>=2.0.0

